refactor(FE): rename and update styles
