+++ frontend/src/layout/Layout.tsx
@@ -1,14 +1,14 @@
 import { Outlet } from 'react-router-dom'
-import Navbar from '../components/NavigationBar'
+import Navbar from '../components/Navbar'
 
 const Layout = () => {
   return (
     <>
-      <div className='h-screen w-screen flex flex-col'>
-        <div className=' h-28 w-full border-gray-800'>
+      <div className='w-screen px-4 flex flex-col'>
+        <div className='h-28 w-full border-gray-800'>
           <Navbar />
         </div>
-        <div className='flex-1 w-full h-full'>
+        <div className='w-full h-full flex items-center justify-center'>
           <Outlet />
         </div>
       </div>
